core: move CDC configuration types into core/cdc.rs
core/lib.rs had grown a 125-line block of CDC-specific types
(CaptureDataChangesMode, CdcVersion, CaptureDataChangesInfo and the
CaptureDataChangesExt trait) sitting between crate wiring and
QueryRunner. That was the only domain code in the file, and it made
lib.rs harder to scan for what it is actually for: module declarations,
re-exports and a few global constants.

Move that block into a new core/cdc.rs module and re-export the types
from the crate root so external users (sync/engine, integration tests)
keep using turso_core::CdcVersion and friends unchanged. Also move the
TURSO_CDC_DEFAULT_TABLE_NAME and TURSO_CDC_VERSION_TABLE_NAME constants
out of translate/pragma.rs into the new module, since they belong with
the rest of the CDC configuration.

No behavior change.

Tests: cargo clippy --workspace --all-features --all-targets, CDC
integration tests (cargo test -p core_tester --test integration_tests
-- cdc).
